refactor: reorganize files and exports
