+++ src/core/entity-state.ts
@@ -1,0 +1,5 @@
+export enum EntityState {
+  ADDED,
+  LOADED,
+  DELETED,
+}
