+++ src/core/repository.ts
@@ -1,37 +1,10 @@
 import { WeakIdentityMap } from 'weak-identity-map';
 import { DBContext } from './db-context';
 import { PersistenceError, PersistenceOperation } from './persistence.error';
-import { serialize } from 'node:v8';
 import * as EventEmitter from 'emittery';
 import { Transaction, TransactionEvents } from './transaction/transaction';
-
-type EntityState = 'added' | 'loaded' | 'deleted';
-
-class EntityWrapper<E extends object> {
-  public readonly entity: E;
-  public state: EntityState;
-  private snapshot: Buffer;
-
-  constructor(entity: E, state: EntityState) {
-    this.entity = entity;
-    this.state = state;
-
-    this.snapshot = this.createSnapshot();
-  }
-
-  verify() {
-    const oldSnapshot = this.snapshot;
-    const newSnapshot = this.createSnapshot();
-
-    this.snapshot = newSnapshot;
-
-    return oldSnapshot.equals(newSnapshot);
-  }
-
-  private createSnapshot() {
-    return serialize(this.entity);
-  }
-}
+import { EntityWrapper } from './entity-wrapper';
+import { EntityState } from './entity-state';
 
 export interface RepositoryConstructor<
   R,
@@ -82,11 +55,11 @@ export abstract class Repository<
   }
 
   add(entity: E): boolean {
-    return this.track(entity, 'added') === entity;
+    return this.track(entity, EntityState.ADDED) === entity;
   }
 
   delete(entity: E) {
-    return this.track(entity, 'deleted') === entity;
+    return this.track(entity, EntityState.DELETED) === entity;
   }
 
   protected trackAll<P extends E | undefined>(
@@ -165,20 +138,20 @@ export abstract class Repository<
           }
         };
 
-        if (wrapper.state === 'added') {
+        if (wrapper.state === EntityState.ADDED) {
           await assertChange(() => this.insert(wrapper.entity), 'insert');
-          wrapper.state = 'loaded';
+          wrapper.state = EntityState.LOADED;
 
           this.emit('inserted', {
             entity: wrapper.entity,
           }).catch(console.error);
-        } else if (wrapper.state === 'loaded' && !wrapper.verify()) {
+        } else if (wrapper.state === EntityState.LOADED && !wrapper.verify()) {
           await assertChange(async () => this.update(wrapper.entity), 'update');
 
           this.emit('updated', {
             entity: wrapper.entity,
           }).catch(console.error);
-        } else if (wrapper.state === 'deleted') {
+        } else if (wrapper.state === EntityState.DELETED) {
           await assertChange(() => this.remove(wrapper.entity), 'delete');
 
           this.identityMap.delete(id);
