--- src/index.ts
@@ -1,30 +1,0 @@
-export { Uow, RunOptions } from './core/unit-of-work';
-export { DBContext } from './core/db-context';
-export { Context } from './core/context';
-export {
-  Repository,
-  RepositoryConstructor,
-  RepositoryEvents,
-} from './core/repository';
-export { DataMapper, DataMapperConstructor } from './core/data-mapper';
-export {
-  Transaction,
-  TransactionEvents,
-  TranscationEventListener,
-} from './core/transaction/transaction';
-export { Engine } from './core/transaction/engine';
-export { PersistenceError } from './core/persistence.error';
-export { RunError } from './core/run-error';
-export { BaseProperty } from './core/orm/base-property';
-export { Property } from './core/orm/property';
-export { JsonProperty } from './core/orm/json-property';
-export { EntityPropertiesMap } from './core/orm/entity-properties-map';
-export { ObjectOperator } from './core/orm/object-operator';
-export { createDataMapper } from './knex/entity-data-mapper';
-// TODO: export everythings
-export { Schema } from './knex/schema';
-export { KnexEngine } from './knex/knex.engine';
-export {
-  KnexTransaction,
-  KnexTransactionOptions,
-} from './knex/knex.transaction';
