+++ src/knex/base-property.ts
