+++ src/knex/entity-properties-map.ts
