+++ src/knex/json-property.ts
