+++ tests/orm/delete-existing-model.spec.ts
@@ -1,10 +1,10 @@
 import { Knex, knex } from 'knex';
 import {
-  KnexEngine,
   KnexTransaction,
   KnexTransactionOptions,
-  Uow,
-} from '../../src';
+  KnexEngine,
+} from '../../src/knex';
+import { Uow } from '../../src/core';
 import { resolve } from 'path';
 import { CustomerRepository } from '../utils/customer.schema';
 import { faker } from '@faker-js/faker/locale/yo_NG';
